Move convert_v3_to_v2_metadata to utils
